myproject/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.1
pytz==2025.1
Pillow==11.1.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0